requirements.txt

transformers==4.45.1
diffusers==0.31.0
accelerate==0.34.2
safetensors==0.4.4
peft>=0.11.0
torch
Pillow
streamlit
